% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-plugins-metadata-utils.R
\name{check_type}
\alias{check_type}
\title{Check that an object is an OMOP-ES plugin}
\usage{
check_type(plugin)
}
\arguments{
\item{plugin}{Object to check}
}
\value{
\code{NULL}, invisibly. Called for its side effect of raising an error.
}
\description{
Errors unless \code{plugin} has \code{"omop_plugin"} among its classes. Used to fail
early, with a clear message, when the contents of \code{omop_plugins} are not
what the introspection functions expect.
}
\details{
This function is copied from OMOP-ES (licence: GPL-3).
}
\seealso{
Other OMOP-ES plugin introspection:
\code{\link[=enabled_by_settings]{enabled_by_settings()}},
\code{\link[=omop_es_plugins_extract_docs_private]{omop_es_plugins_extract_docs_private()}},
\code{\link[=omop_es_plugins_extract_docs_public]{omop_es_plugins_extract_docs_public()}},
\code{\link[=omop_es_plugins_extract_metadata]{omop_es_plugins_extract_metadata()}},
\code{\link[=omop_es_plugins_extract_sql]{omop_es_plugins_extract_sql()}},
\code{\link[=omop_es_plugins_extract_tables]{omop_es_plugins_extract_tables()}},
\code{\link[=plugin_extract_sql]{plugin_extract_sql()}},
\code{\link[=plugin_extract_tables]{plugin_extract_tables()}},
\code{\link[=plugins_extract_sql]{plugins_extract_sql()}},
\code{\link[=plugins_extract_tables]{plugins_extract_tables()}},
\code{\link[=read_table_level_md]{read_table_level_md()}},
\code{\link[=read_table_level_private_md]{read_table_level_private_md()}}
}
\concept{OMOP-ES plugin introspection}
\keyword{internal}
