% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dbplyr.R
\name{db_from_tbl}
\alias{db_from_tbl}
\title{Database connection underlying a lazy table}
\usage{
db_from_tbl(tbl)
}
\arguments{
\item{tbl}{A lazy \code{tbl} with a database source}
}
\value{
A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object.
}
\description{
Extracts the \link[DBI:DBIConnection-class]{DBI::DBIConnection} that a \pkg{dbplyr} lazy table is
querying, so that further SQL can be issued against the same database.
}
\keyword{internal}
