% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-schema-tools.R
\name{dbListTablesAndViewsInSchema}
\alias{dbListTablesAndViewsInSchema}
\title{Get all tables and views in a schema}
\usage{
dbListTablesAndViewsInSchema(conn, schema)
}
\arguments{
\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{schema}{The schema name, a character string}
}
\value{
A character vector of table or view names
}
\description{
Lists both base tables and views in \code{schema}, by querying
\code{information_schema.tables}.
}
\seealso{
Other database schema helpers:
\code{\link[=dbCreateSchema]{dbCreateSchema()}},
\code{\link[=dbListOmopTables]{dbListOmopTables()}},
\code{\link[=dbListTablesInSchema]{dbListTablesInSchema()}}
}
\concept{database schema helpers}
\keyword{internal}
