% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-register.R
\name{duckdb_register_omop_es_output}
\alias{duckdb_register_omop_es_output}
\title{Register OMOP-ES datalake as duckdb table}
\usage{
duckdb_register_omop_es_output(con, extract_path, omop_es_path)
}
\arguments{
\item{con}{A database connection}

\item{extract_path}{Path to folder containing OMOP-ES extract}

\item{omop_es_path}{Path to OMOP-ES directory (used for registering
concept tables from the `omop_metadata` directory)}
}
\description{
A single extract from OMOP-ES produces one of two formats:
}
\details{
1. A folder of CSV files
2. A new (timestampped) directory as a subdirectory of `extract` e.g.
  `extract/CUH_EPIC_batch_cohort-20260201_090000`.
  Within this, there are 3 directories (`public`, `private` and `custom`),
  each of which contains a directory for each OMOP table (e.g.
  `condition_occurrence`), which then contains multiple `*.parquet` files.

This function registers:
1. Tables under `public` in the `dbo` schema
2. Tables under `private` in the `priv` schema
3. Tables under `custom` in the `dbo` schema
4. Tables in the `omop_metadata` subdirectory of `omop_es_path` in the `dbo`
   schema
}
