% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_is_clean}
\alias{git_is_clean}
\title{Is a git repository clean?}
\usage{
git_is_clean(repo)
}
\arguments{
\item{repo}{Path to a git repository}
}
\value{
\code{TRUE} if the repository is clean, otherwise \code{FALSE}.
}
\description{
Whether a git repository has no changes reported by \code{\link[gert:git_status]{gert::git_status()}},
i.e. no staged, unstaged or untracked changes.
}
\keyword{internal}
