% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-linking.R
\name{is_col_class_same}
\alias{is_col_class_same}
\title{Do two tables agree on the class of a column?}
\usage{
is_col_class_same(table1, table2, column)
}
\arguments{
\item{table1, table2}{Lazy \code{tbl} objects, each containing \code{column}}

\item{column}{Name of the column to compare, a character string}
}
\value{
\code{TRUE} if the classes are identical, otherwise \code{FALSE}.
}
\description{
Compares the R class that a column is materialised as in each of two lazy
tables, by collecting a single row from each. Mismatched types can stop
joins between the tables from working.
}
\keyword{internal}
