% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-linking.R
\name{omop_es_link_tables_for_foreign_key_columns}
\alias{omop_es_link_tables_for_foreign_key_columns}
\title{OMOP-ES links tables for an OMOP table's key columns}
\usage{
omop_es_link_tables_for_foreign_key_columns(table)
}
\arguments{
\item{table}{OMOP table name}
}
\value{
A character vector (strictly, a \pkg{glue} vector) of \verb{_links}
table names.
}
\description{
The names of the OMOP-ES \verb{_links} tables that relate to \code{table}, that is
one per table returned by \code{\link[=omop_source_tables_for_foreign_key_columns]{omop_source_tables_for_foreign_key_columns()}}.
}
\seealso{
\code{\link[=omop_es_tbl_with_links]{omop_es_tbl_with_links()}}, which joins these tables on.
}
\keyword{internal}
