% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-markdown.R
\name{read_table_level_md}
\alias{read_table_level_md}
\title{Load table-level public documentation Markdown files}
\usage{
read_table_level_md(table_name, omop_es_path)
}
\arguments{
\item{omop_es_path}{Path to OMOP-ES directory}

\item{table}{The OMOP table e.g. `"condition_occurrence"`}
}
\value{
Character vector of length 1 containing the Markdown file
}
\description{
Loads the documentation stored in `docs/CUH/{table_name}.md`
within the OMOP-ES directory `omop_es_path`.
}
