% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-tbl.R
\name{tbl_omop}
\alias{tbl_omop}
\title{Lazy table for an OMOP table}
\usage{
tbl_omop(conn, table_name, schema = "dbo")
}
\arguments{
\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{table_name}{OMOP table name, e.g. \code{"condition_occurrence"}}

\item{schema}{Name of the schema holding the public OMOP tables}
}
\value{
A lazy \code{tbl}.
}
\description{
A \code{\link[dplyr:tbl]{dplyr::tbl()}} for an OMOP table in the given schema.
}
\seealso{
Other OMOP table references:
\code{\link[=id_omop]{id_omop()}},
\code{\link[=id_omop_links]{id_omop_links()}},
\code{\link[=tbl_omop_concept]{tbl_omop_concept()}},
\code{\link[=tbl_omop_links]{tbl_omop_links()}}
}
\concept{OMOP table references}
\keyword{internal}
