% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-tbl.R
\name{tbl_omop_concept}
\alias{tbl_omop_concept}
\title{Lazy table for the OMOP concept table}
\usage{
tbl_omop_concept(conn, mapping_date = Sys.Date())
}
\arguments{
\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{mapping_date}{The date at which validity of OMOP concepts should be assessed}
}
\value{
A lazy \code{tbl}.
}
\description{
A \code{\link[dplyr:tbl]{dplyr::tbl()}} for the OMOP vocabulary \code{concept} table. Note that the
schema is hard-coded to \code{"dbo"}.
}
\seealso{
\code{\link[=decorate_mapping_table]{decorate_mapping_table()}}, which uses this to annotate concept
ids.

Other OMOP table references:
\code{\link[=id_omop]{id_omop()}},
\code{\link[=id_omop_links]{id_omop_links()}},
\code{\link[=tbl_omop]{tbl_omop()}},
\code{\link[=tbl_omop_links]{tbl_omop_links()}}
}
\concept{OMOP table references}
\keyword{internal}
