% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-tbl.R
\name{tbl_omop_links}
\alias{tbl_omop_links}
\title{Lazy table for an OMOP-ES links table}
\usage{
tbl_omop_links(conn, table_name, schema = "priv")
}
\arguments{
\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{table_name}{OMOP table name, e.g. \code{"condition_occurrence"}}

\item{schema}{Name of the schema holding the private OMOP-ES tables}
}
\value{
A lazy \code{tbl}.
}
\description{
A \code{\link[dplyr:tbl]{dplyr::tbl()}} for the OMOP-ES \verb{_links} table belonging to an OMOP
table.
}
\seealso{
Other OMOP table references:
\code{\link[=id_omop]{id_omop()}},
\code{\link[=id_omop_links]{id_omop_links()}},
\code{\link[=tbl_omop]{tbl_omop()}},
\code{\link[=tbl_omop_concept]{tbl_omop_concept()}}
}
\concept{OMOP table references}
\keyword{internal}
