% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-schema-tools.R
\name{dbCreateSchema}
\alias{dbCreateSchema}
\title{Create or replace schema in a database}
\usage{
dbCreateSchema(conn, schema)
}
\arguments{
\item{conn}{A [DBI::DBIConnection] object, as returned by [DBI::dbConnect()].}

\item{schema}{The schema name, a character string}
}
\description{
Create or replace schema in a database
}
