% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-schema-tools.R
\name{dbListTablesAndViewsInSchema}
\alias{dbListTablesAndViewsInSchema}
\title{Get all tables and views in a schema}
\usage{
dbListTablesAndViewsInSchema(conn, schema)
}
\arguments{
\item{conn}{A [DBI::DBIConnection] object, as returned by [DBI::dbConnect()].}

\item{schema}{The schema name, a character string}
}
\value{
A character vector of table or view names
}
\description{
Get all tables and views in a schema
}
