% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-schema-tools.R
\name{dbListTablesInSchema}
\alias{dbListTablesInSchema}
\title{List all tables in a schema}
\usage{
dbListTablesInSchema(conn, schema)
}
\arguments{
\item{conn}{A [DBI::DBIConnection] object, as returned by [DBI::dbConnect()].}

\item{schema}{The schema name, a character string}
}
\value{
A character vector of table names
}
\description{
List all tables in a schema
}
