% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-register.R
\name{duckdb_register_omop_es_datalake}
\alias{duckdb_register_omop_es_datalake}
\title{Register a data-lake parquet files as duckdb table}
\usage{
duckdb_register_omop_es_datalake(con, folder_path, schema = NULL)
}
\arguments{
\item{con}{A database connection}

\item{folder_path}{Path to folder containing parquet files (one parquet per
table)}

\item{schema}{Name of schema to register tables in}
}
\description{
Given a folder containing one folder per table, with each containing
several parquet files, this function registers these as tables in a duckdb
database
}
