% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-register.R
\name{duckdb_register_parquet_dir}
\alias{duckdb_register_parquet_dir}
\title{Register a folder of parquet files as duckdb table}
\usage{
duckdb_register_parquet_dir(con, folder_path, schema = NULL)
}
\arguments{
\item{con}{A database connection}

\item{folder_path}{Path to folder containing parquet files (one parquet per
table)}

\item{schema}{Name of schema to register tables in}
}
\description{
Register a folder of parquet files as duckdb table
}
