% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-local.R
\name{omop_es_diff_viewer_local_git}
\alias{omop_es_diff_viewer_local_git}
\title{Run OMOP-ES for two git SHAs and compare}
\usage{
omop_es_diff_viewer_local_git(
  omop_es_path,
  left_branch,
  right_branch,
  settings_id = "CUH_EPIC_small_cohort",
  cohort_limit = 5000,
  pull = TRUE,
  envvar = callr::rcmd_safe_env()
)
}
\arguments{
\item{omop_es_path}{Path to OMOP-ES directory}

\item{left_branch}{A git branch or SHA}

\item{right_branch}{A git branch or SHA}

\item{settings_id}{The OMOP-ES settings to use}

\item{cohort_limit}{The max number of patients to use.}

\item{envvar}{A list of environment variables to set in the child process
prior to running the pipeline. This can be used to pass e.g. database
connection details to OMOP-ES}
}
\description{
Run OMOP-ES for two git SHAs and compare
}
