% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-mappers.R
\name{omop_es_plugins_extract_docs_private}
\alias{omop_es_plugins_extract_docs_private}
\title{Load all table-level private documentation Markdown files}
\usage{
omop_es_plugins_extract_docs_private(
  omop_es_path,
  settings_id = "CUH_EPIC_small_cohort",
  cohort_limit = 10
)
}
\arguments{
\item{omop_es_path}{Path to OMOP-ES directory}

\item{settings_id}{The OMOP-ES settings to use}

\item{cohort_limit}{The max number of patients to use. This needs to be
small enough to be fast, but large enough to avoid odd quirks (e.g. none
of the included patients have an imaging results)}
}
\value{
A named list, with one element per OMOP table. Each element of the
  list contains the corresponding Markdown code
}
\description{
Loads the documentation stored in `docs/CUH/{table_name}_private.md`
within the OMOP-ES directory `omop_es_path` for every OMOP table with a
mapper specified.
}
