% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-mappers.R
\name{omop_es_plugins_extract_tables}
\alias{omop_es_plugins_extract_tables}
\title{Extract database tables used by each OMOP-ES plugin}
\usage{
omop_es_plugins_extract_tables(
  omop_es_path,
  settings_id = "CUH_EPIC_small_cohort",
  cohort_limit = 10
)
}
\arguments{
\item{omop_es_path}{Path to OMOP-ES directory}

\item{settings_id}{The OMOP-ES settings to use}

\item{cohort_limit}{The max number of patients to use. This needs to be
small enough to be fast, but large enough to avoid odd quirks (e.g. none
of the included patients have an imaging results)}
}
\value{
A named nested list of character database table names The outer
  named list contains one element for each OMOP table. Within each
  table-level element, there is a named list containing one element per
  plugin.
}
\description{
For every plug-in in the supplied OMOP-ES directory (`omop_es_path`), this
function runs the plugin on the supplied cohort, but overrides the
relevant R functions that query the database (`tbl()`) so that we can
extract the database tables that the plug-in uses.
}
