% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-run.R
\name{omop_es_run}
\alias{omop_es_run}
\title{Run OMOP-ES in a separate process}
\usage{
omop_es_run(
  omop_es_path,
  settings_id = "CUH_EPIC_small_cohort",
  cohort_limit = 5000,
  output_parquet = NA,
  zip_output = FALSE,
  custom_dir = NULL,
  envvar = callr::rcmd_safe_env()
)
}
\arguments{
\item{omop_es_path}{Path to OMOP-ES directory}

\item{settings_id}{The OMOP-ES settings to use}

\item{cohort_limit}{The max number of patients to use.}

\item{zip_output}{Whether to zip output}

\item{custom_dir}{A custom directory}

\item{envvar}{A list of environment variables to set in the child process
prior to running the pipeline. This can be used to pass e.g. database
connection details to OMOP-ES}
}
\description{
Runs an OMOP-ES pipeline in a separate process so that it isolated from the
the current state of the R environment.
}
