% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-es-linking.R
\name{omop_es_tbl_with_links}
\alias{omop_es_tbl_with_links}
\title{Join OMOP-ES table to links tables}
\usage{
omop_es_tbl_with_links(
  conn,
  table,
  schema_public = "dbo",
  schema_private = "priv",
  drop_omop_foreign_keys = FALSE
)
}
\description{
Join OMOP-ES table to links tables
}
