% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-metadata.R
\name{omop_metadata_table_level}
\alias{omop_metadata_table_level}
\title{Load the OMOP CDM table level definition table}
\usage{
omop_metadata_table_level()
}
\description{
Load the OMOP CDM table level definition table
}
