% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-metadata.R
\name{omop_source_tables_for_foreign_key_columns}
\alias{omop_source_tables_for_foreign_key_columns}
\title{Foreign key column names for OMOP table}
\usage{
omop_source_tables_for_foreign_key_columns(table)
}
\arguments{
\item{table}{OMOP table name}
}
\description{
Foreign key column names for OMOP table
}
