% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-metadata.R
\name{omop_table_concept_columns}
\alias{omop_table_concept_columns}
\title{Extract the stubs of *_concept_id columns}
\usage{
omop_table_concept_columns(db, table)
}
\description{
Extract the stubs of *_concept_id columns
}
