% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-metadata.R
\name{omop_vocab_tables}
\alias{omop_vocab_tables}
\title{Names of the OMOP CDM vocabulary tables}
\usage{
omop_vocab_tables()
}
\value{
A character vector of OMOP table names.
}
\description{
The tables in the OMOP CDM v5.4 specification whose \code{schema} is \code{"VOCAB"},
i.e. the standardised vocabulary tables such as \code{concept} and
\code{concept_relationship}.
}
\seealso{
\code{\link[=dbListOmopTables]{dbListOmopTables()}}, which uses this to optionally exclude
vocabulary tables from a listing.

Other OMOP CDM metadata:
\code{\link[=omop_all_tables]{omop_all_tables()}},
\code{\link[=omop_cdm_tables]{omop_cdm_tables()}},
\code{\link[=omop_metadata_field_level]{omop_metadata_field_level()}},
\code{\link[=omop_metadata_table_level]{omop_metadata_table_level()}},
\code{\link[=omop_source_tables_for_foreign_key_columns]{omop_source_tables_for_foreign_key_columns()}},
\code{\link[=omop_table_all_key_columns]{omop_table_all_key_columns()}},
\code{\link[=omop_table_columns]{omop_table_columns()}},
\code{\link[=omop_table_common_columns]{omop_table_common_columns()}},
\code{\link[=omop_table_primary_key]{omop_table_primary_key()}}
}
\concept{OMOP CDM metadata}
\keyword{internal}
