% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-plugins-extract-info.R
\name{plugin_extract_sql}
\alias{plugin_extract_sql}
\title{Extract SQL queries used in a single OMOP-ES plugin}
\usage{
plugin_extract_sql(plugin, name, conns, cohort)
}
\arguments{
\item{plugin}{An OMOP-ES plugin function}

\item{name}{The name of the OMOP-ES plugin}

\item{conns}{The OMOP-ES `conns` object (a list of database connections)}

\item{cohort}{The OMOP-ES `cohort` tibble}
}
\value{
A list of character SQL queries
}
\description{
The function runs the plugin on the supplied cohort, but overrides the
relevant R functions that query the database (`collect()` and
`dbGetQuery()`) so that we can extract the SQL queries that the plug-in
uses.
}
