% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop-plugins-extract-info.R
\name{plugins_extract_tables}
\alias{plugins_extract_tables}
\title{Extract all database tables used in all OMOP-ES plugins}
\usage{
plugins_extract_tables(omop_plugins, conns, cohort)
}
\arguments{
\item{omop_plugins}{A list of OMOP-ES plugins}

\item{conns}{The OMOP-ES `conns` object (a list of database connections)}

\item{cohort}{The OMOP-ES `cohort` tibble}
}
\value{
A nested list of list of database table names
}
\description{
The function runs all the `omop_plugins` on the supplied cohort, but
overrides the relevant R functions that query the database (`collect()` and
`dbGetQuery()`) so that we can extract database tables that the plug-in
uses.
}
